+++ src/github.com/stellar/horizon/middleware_logger.go
@@ -2,12 +2,14 @@ package horizon
 
 import (
 	"net/http"
+	"strings"
 	"time"
 
 	"golang.org/x/net/context"
 
 	gctx "github.com/goji/context"
 	"github.com/stellar/horizon/log"
+	"github.com/stellar/horizon/render"
 	"github.com/zenazn/goji/web"
 	"github.com/zenazn/goji/web/middleware"
 	"github.com/zenazn/goji/web/mutil"
@@ -30,8 +32,11 @@ func LoggerMiddleware(c *web.C, h http.Handler) http.Handler {
 		then := time.Now()
 		h.ServeHTTP(mw, r)
 		duration := time.Now().Sub(then)
-
-		logEndOfRequest(ctx, duration, mw)
+		// Checking `Accept` header from user request because if the streaming connection
+		// is reset before sending the first event no Content-Type header is sent in a response.
+		acceptHeader := r.Header.Get("Accept")
+		streaming := strings.Contains(acceptHeader, render.MimeEventStream)
+		logEndOfRequest(ctx, r, duration, mw, streaming)
 	}
 
 	return http.HandlerFunc(fn)
@@ -46,10 +51,15 @@ func logStartOfRequest(ctx context.Context, r *http.Request) {
 	}).Info("Starting request")
 }
 
-func logEndOfRequest(ctx context.Context, duration time.Duration, mw mutil.WriterProxy) {
+func logEndOfRequest(ctx context.Context, r *http.Request, duration time.Duration, mw mutil.WriterProxy, streaming bool) {
 	log.Ctx(ctx).WithFields(log.F{
-		"status":   mw.Status(),
-		"bytes":    mw.BytesWritten(),
-		"duration": duration,
+		"path":      r.URL.String(),
+		"method":    r.Method,
+		"ip":        r.RemoteAddr,
+		"host":      r.Host,
+		"status":    mw.Status(),
+		"bytes":     mw.BytesWritten(),
+		"duration":  duration,
+		"streaming": streaming,
 	}).Info("Finished request")
 }